# docs/cust_auth_swagger_api.yaml
openapi: 3.0.3
info:
  title: Banking Technology Lab — Auth & Customer Module API
  version: '1.0.0'
  description: |
    Covers two modules of the Banking Technology Lab (CSF 3142) project:

    * **Login / Auth** — registration, login, JWT issuance, role-based access
    * **Customer Management** — customer profile plus its typed child records
      (Name, Address, Contact Info, Proof of ID, Identification) and the
      shared Classification lookup table

    Roles: `CUSTOMER`, `BANK_OFFICER`, `ADMIN`. The required role for each
    endpoint is stated explicitly in its description — `BANK_OFFICER` and
    `ADMIN` can generally act on any customer; `CUSTOMER` can only act on
    their own record, resolved from the JWT, never from a path/query param.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Public endpoint, no authentication required.
        Hashes the password with BCrypt before storage; rejects duplicate usernames.
      security: []
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error (missing/malformed fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive a JWT
      description: |
        Public endpoint, no authentication required.
        Validates credentials against the stored BCrypt hash and returns a
        JWT containing the user's role claim on success.
      security: []
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/users:
    post:
      tags: [Auth]
      summary: Create a BANK_OFFICER or ADMIN account (admin-provisioned)
      description: |
        **Required role:** `ADMIN`.
        Not public — requires `Authorization: Bearer <JWT>` for an existing
        ADMIN. Used to provision privileged accounts once the very first
        ADMIN exists (see the project README for how that first ADMIN is
        bootstrapped). The "role" in the request body only ever describes
        the NEW user being created — the caller's own privilege comes
        entirely from their validated JWT, never from this request body.
        `CUSTOMER` cannot be created here; customers self-register via
        `POST /api/auth/register`.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error, or role is not BANK_OFFICER/ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Caller is authenticated but not an ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/api/auth/login`, containing the user's role claim.
        Sent as `Authorization: Bearer <token>`.
...
  schemas:
    Role:
      type: string
      enum: [CUSTOMER, BANK_OFFICER, ADMIN]

    # ---------------- Auth ----------------
    RegisterRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: jane_doe
        password:
          type: string
          format: password
          minLength: 8
          example: correct-horse-battery-staple
        email:
          type: string
          format: email
          example: jane.doe@example.com
    RegisterResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        message:
          type: string
          example: User registered successfully
    CreateUserRequest:
      type: object
      required: [username, password, email, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: officer01
        email:
          type: string
          format: email
          example: officer@bank.com
        password:
          type: string
          format: password
          minLength: 8
          example: Bank@123
        role:
          type: string
          enum: [BANK_OFFICER, ADMIN]
          description: Role to assign to the new user. CUSTOMER cannot be created through this endpoint.
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Seconds until expiry
          example: 36000
        role:
          $ref: '#/components/schemas/Role'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
        code:
          type: string
          nullable: true