# src/main/java/com/example/demo/dto/LoginResponse.java
package com.example.demo.dto;

public class LoginResponse {

    private String token;
    private String tokenType;
    private long expiresIn;
    private String role;

    public LoginResponse() {
    }

    public LoginResponse(String token, String tokenType, long expiresIn, String role) {
        this.token = token;
        this.tokenType = tokenType;
        this.expiresIn = expiresIn;
        this.role = role;
    }

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }

    public String getTokenType() {
        return tokenType;
    }

    public void setTokenType(String tokenType) {
        this.tokenType = tokenType;
    }

    public long getExpiresIn() {
        return expiresIn;
    }

    public void setExpiresIn(long expiresIn) {
        this.expiresIn = expiresIn;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }
}